chore: Move ConversationFailedToAddParticipantsSystemMessageCellDescription into its own file
