+++ wire-ios/Wire-iOS/Sources/UserInterface/Conversation/Content/Cells/ConfigurationMessageCell/Components/ConversationSystemMessageCell.swift
@@ -193,57 +193,3 @@ private extension ConversationLike {
 }
 
 // MARK: - Descriptions
-
-final class ConversationFailedToAddParticipantsSystemMessageCellDescription: ConversationMessageCellDescription {
-
-    typealias SystemContent = L10n.Localizable.Content.System
-    typealias View = FailedUsersSystemMessageCell
-
-    let configuration: View.Configuration
-
-    var message: ZMConversationMessage?
-    weak var delegate: ConversationMessageCellDelegate?
-    weak var actionController: ConversationMessageActionController?
-
-    var showEphemeralTimer: Bool = false
-    var topMargin: Float = 26.0
-
-    let isFullWidth: Bool = true
-    let supportsActions: Bool = false
-    let containsHighlightableContent: Bool = false
-
-    let accessibilityIdentifier: String? = nil
-    let accessibilityLabel: String? = nil
-
-    init(failedUsers: [UserType], isCollapsed: Bool, buttonAction: @escaping Completion) {
-        configuration = View.Configuration(
-            title: ConversationFailedToAddParticipantsSystemMessageCellDescription.configureTitle(for: failedUsers),
-            content: ConversationFailedToAddParticipantsSystemMessageCellDescription.configureContent(for: failedUsers),
-            isCollapsed: isCollapsed,
-            icon: Asset.Images.attention.image,
-            buttonAction: buttonAction)
-    }
-
-    private static func configureTitle(for failedUsers: [UserType]) -> NSAttributedString? {
-        guard failedUsers.count > 1 else {
-            return nil
-        }
-
-        let title = SystemContent.FailedtoaddParticipants.count(failedUsers.count)
-        return .markdown(from: title, style: .errorLabelStyle)
-    }
-
-    private static func configureContent(for failedUsers: [UserType]) -> NSAttributedString {
-        let keyString = "content.system.failedtoadd_participants.could_not_be_added"
-
-        let userNames = failedUsers.compactMap { $0.name }
-        let userNamesJoined = userNames.joined(separator: ", ")
-        let text = keyString.localized(args: userNames.count, userNamesJoined)
-
-        let attributedText = NSAttributedString.errorSystemMessage(withText: text, andHighlighted: userNamesJoined)
-        let learnMore = NSAttributedString.unreachableBackendLearnMoreLink
-
-        return [attributedText, learnMore].joined(separator: " ".attributedString)
-    }
-
-}
